feat: unified `polykey.ts` and `polykey-agent.ts` so forking to background is portable
